Update task service

+++ src/modules/tasks/tasks.service.ts
@@ -1,18 +1,21 @@
 import { FastifyRequest } from 'fastify';
-import { eq } from 'drizzle-orm';
+import { and, eq } from 'drizzle-orm';
 
 import { tasks } from '@db/schema';
 
 export const getTasks = async (request: FastifyRequest) => {
   const { db } = request.server;
 
-  return await db.select().from(tasks);
+  return await db.select().from(tasks).where(eq(tasks.userId, request.user.id));
 };
 
 export const getTaskById = async (request: FastifyRequest, id: number) => {
   const { db } = request.server;
 
-  const [task] = await db.select().from(tasks).where(eq(tasks.id, id));
+  const [task] = await db
+    .select()
+    .from(tasks)
+    .where(and(eq(tasks.id, id), eq(tasks.userId, request.user.id)));
 
   return task;
 };
@@ -23,7 +26,10 @@ export const createTask = async (
 ) => {
   const { db } = request.server;
 
-  const [task] = await db.insert(tasks).values(data).returning();
+  const [task] = await db
+    .insert(tasks)
+    .values({ ...data, userId: request.user.id })
+    .returning();
 
   return getTaskById(request, task.id);
 };
@@ -34,6 +40,15 @@ export const updateTask = async (
 ) => {
   const { db } = request.server;
 
+  const [task] = await db
+    .select()
+    .from(tasks)
+    .where(and(eq(tasks.id, data.id), eq(tasks.userId, request.user.id)));
+
+  if (!task) {
+    throw new Error('Task not found');
+  }
+
   await db.update(tasks).set(data).where(eq(tasks.id, data.id));
 
   return getTaskById(request, data.id);
@@ -41,6 +56,15 @@ export const updateTask = async (
 
 export const deleteTask = async (request: FastifyRequest, id: number) => {
   const { db } = request.server;
+
+  const [task] = await db
+    .select()
+    .from(tasks)
+    .where(and(eq(tasks.id, id), eq(tasks.userId, request.user.id)));
+
+  if (!task) {
+    throw new Error('Task not found');
+  }
 
   await db.delete(tasks).where(eq(tasks.id, id));
 };
